backend/requirements/base.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
pypdf==5.4.0
